streamingccl: move some functions into the files they are used in
Its a little nicer when functions are defined closer to where they are
used.

Epic: none

Release note: None
