+++ pkg/ccl/streamingccl/streamingest/stream_ingest_manager.go
@@ -17,6 +17,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
 	"github.com/cockroachdb/cockroach/pkg/repstream"
 	"github.com/cockroachdb/cockroach/pkg/repstream/streampb"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/sql"
 	"github.com/cockroachdb/cockroach/pkg/sql/isql"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
@@ -25,6 +26,8 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
 	"github.com/cockroachdb/cockroach/pkg/sql/syntheticprivilege"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/cockroachdb/errors"
 )
 
 type streamIngestManagerImpl struct {
@@ -83,6 +86,67 @@ func newStreamIngestManagerWithPrivilegesCheck(
 		txn:         txn,
 		jobRegistry: execCfg.JobRegistry,
 	}, nil
+}
+
+// applyCutoverTime modifies the consumer job record with a cutover time and
+// unpauses the job if necessary.
+func applyCutoverTime(
+	ctx context.Context,
+	jobRegistry *jobs.Registry,
+	txn isql.Txn,
+	ingestionJobID jobspb.JobID,
+	cutoverTimestamp hlc.Timestamp,
+) error {
+	log.Infof(ctx, "adding cutover time %s to job record", cutoverTimestamp)
+	if err := jobRegistry.UpdateJobWithTxn(ctx, ingestionJobID, txn, false,
+		func(txn isql.Txn, md jobs.JobMetadata, ju *jobs.JobUpdater) error {
+			progress := md.Progress.GetStreamIngest()
+			details := md.Payload.GetStreamIngestion()
+			if progress.ReplicationStatus == jobspb.ReplicationCuttingOver {
+				return errors.Newf("job %d already started cutting over to timestamp %s",
+					ingestionJobID, progress.CutoverTime)
+			}
+
+			progress.ReplicationStatus = jobspb.ReplicationPendingCutover
+			// Update the sentinel being polled by the stream ingestion job to
+			// check if a complete has been signaled.
+			progress.CutoverTime = cutoverTimestamp
+			progress.RemainingCutoverSpans = roachpb.Spans{details.Span}
+			ju.UpdateProgress(md.Progress)
+			return nil
+		}); err != nil {
+		return err
+	}
+	// Unpause the job if it is paused.
+	return jobRegistry.Unpause(ctx, txn, ingestionJobID)
+}
+
+func getReplicationStatsAndStatus(
+	ctx context.Context, jobRegistry *jobs.Registry, txn isql.Txn, ingestionJobID jobspb.JobID,
+) (*streampb.StreamIngestionStats, string, error) {
+	job, err := jobRegistry.LoadJobWithTxn(ctx, ingestionJobID, txn)
+	if err != nil {
+		return nil, jobspb.ReplicationError.String(), err
+	}
+	details, ok := job.Details().(jobspb.StreamIngestionDetails)
+	if !ok {
+		return nil, jobspb.ReplicationError.String(),
+			errors.Newf("job with id %d is not a stream ingestion job", job.ID())
+	}
+
+	details.StreamAddress, err = redactSourceURI(details.StreamAddress)
+	if err != nil {
+		return nil, jobspb.ReplicationError.String(), err
+	}
+
+	stats, err := replicationutils.GetStreamIngestionStatsNoHeartbeat(ctx, details, job.Progress())
+	if err != nil {
+		return nil, jobspb.ReplicationError.String(), err
+	}
+	if job.Status() == jobs.StatusPaused {
+		return stats, jobspb.ReplicationPaused.String(), nil
+	}
+	return stats, stats.IngestionProgress.ReplicationStatus.String(), nil
 }
 
 func init() {
